Major refactor of scene module
