+++ build.rs
@@ -67,5 +67,3 @@ fn main() -> io::Result<()> {
 
     Ok(())
 }
-
-// End of File
