+++ src/most_recent.rs
@@ -98,5 +98,3 @@ impl<T> Receiver<T> {
     //     }
     // }
 }
-
-// End of File
