+++ src/scene/event.rs
@@ -1,3 +1,9 @@
+// Copyright (c) 2020 by Lachlan Sneff <lachlan@charted.space>
+// Copyright (c) 2020 by Mark Friedenbach <mark@friedenbach.org>
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
 use std::convert::TryFrom;
 use winit::{
     dpi::{PhysicalPosition, PhysicalSize},
@@ -8,12 +14,6 @@ use winit::{
 pub struct Resize {
     pub new_texture: wgpu::Texture,
     pub size: PhysicalSize<u32>,
-}
-
-#[derive(Debug, Default)]
-pub struct Events {
-    pub resize: Option<Resize>,
-    pub events: Vec<Event>,
 }
 
 #[derive(Debug)]
