+++ src/shaders/icosphere.vert
@@ -1,7 +1,7 @@
 #version 450
 
 layout(set = 0, binding = 0) uniform Globals {
-    mat4 u_Transform;
+    mat4 u_World;
 };
 
 layout(location = 0) in vec3 a_Pos;
@@ -11,5 +11,5 @@ layout(location = 0) out vec4 v_Normal;
 
 void main() {
     v_Normal = vec4(a_Normal, 0.0);
-    gl_Position = u_Transform * vec4(a_Pos, 1.0);
+    gl_Position = u_World * vec4(a_Pos, 1.0);
 }
