+++ src/ui.rs
@@ -7,5 +7,3 @@
 mod root;
 
 pub use root::Root;
-
-// End of File
