+++ src/ui/root.rs
@@ -33,5 +33,3 @@ impl Root {
             .into()
     }
 }
-
-// End of File
